+++ plugins/global-components.js
@@ -1,10 +1,12 @@
 import Vue from 'vue'
+import Box from '@/components/ui/Box/index.vue'
 import Button from '@/components/ui/Button/index.vue'
 import Input from '@/components/ui/Input/index.vue'
 import Label from '@/components/ui/Label/index.vue'
 import Radio from '@/components/ui/Radio/index.vue'
 import Select from '@/components/ui/Select/index.vue'
 import Spinner from '@/components/ui/Spinner/index.vue'
+Vue.component('Box', Box)
 Vue.component('Button', Button)
 Vue.component('Input', Input)
 Vue.component('Label', Label)