+++ test/Index.spec.js
@@ -1,9 +1,14 @@
-import { mount } from '@vue/test-utils'
+import { mount, createLocalVue } from '@vue/test-utils'
+import Box from '../components/ui/Box/index.vue'
 import Index from '@/pages/index.vue'
+
+const localVue = createLocalVue()
+
+localVue.component('Box', Box)
 
 describe('Index', () => {
   test('is a Vue instance', () => {
-    const wrapper = mount(Index)
+    const wrapper = mount(Index, { localVue })
     expect(wrapper.isVueInstance()).toBeTruthy()
   })
   /** Events */
@@ -13,19 +18,19 @@ describe('Index', () => {
     })
     describe('have a HTML button what is clickable', () => {
       test('and trigger `handleDocumentation`', () => {
-        const wrapper = mount(Index)
+        const wrapper = mount(Index, { localVue })
         global.open = jest.fn()
         wrapper.vm.handleDocumentation()
         expect(global.open).toBeCalled()
       })
       test('and trigger `handleGithub`', () => {
-        const wrapper = mount(Index)
+        const wrapper = mount(Index, { localVue })
         global.open = jest.fn()
         wrapper.vm.handleGithub()
         expect(global.open).toBeCalled()
       })
       test('and trigger `handleChangelog`', () => {
-        const wrapper = mount(Index)
+        const wrapper = mount(Index, { localVue })
         global.open = jest.fn()
         wrapper.vm.handleChangelog()
         expect(global.open).toBeCalled()